Removed digit processing from calc, moved it to UI.
Problems with entering negative numbers.

+++ src/main.rs
@@ -5,7 +5,7 @@ mod calc;
 mod ui;
 
 fn main() {
-    let mut calc = Calculator::<f32>::new();
+    let mut calc = Calculator::<f64>::new();
     let mut u = ui::UI::new(&mut calc);
     u.run();
 }
